requirements.txt
# Core LLM libraries
google-generativeai>=0.8.0    # Gemini API
httpx>=0.27.0                  # Modern HTTP client (for API calls)
python-dotenv>=1.0.0           # Load API keys from .env file


langchain-ollama
langgraph>=0.2.0
langchain-google-genai>=2.0.0
langchain-core>=0.3.0
# mcp>=1.0.0
# fastmcp>=0.1.0

ipython>=7.0.0                    # For displaying images in Jupyter Notebooks